+++ rules/deprecated/deprecated.js
@@ -1,0 +1,22 @@
+/**
+ * Deprecated rules listed in reverse chronological order relative to their
+ * deprecation date and annotated with the following information:
+ * package - version when deprecated - replacement rule if any
+ */
+
+module.exports = {
+  /* eslint-disable sort-keys */
+  rules: {
+    'require-jsdoc': 'off', // ESLint - v5.10.0
+    'valid-jsdoc': 'off', // ESLint - v5.10.0
+    'no-catch-shadow': 'off', // ESLint - v5.1.0 - no-shadow
+    'lines-around-directive': 'off', // ESLint - v4.0.0 - padding-line-between-statements
+    'newline-after-var': 'off', // ESLint - v4.0.0 - padding-line-between-statements
+    'newline-before-return': 'off', // ESLint - v4.0.0 - padding-line-between-statements
+    'prefer-reflect': 'off', // ESLint - v3.9.0
+    'no-native-reassign': 'off', // ESLint - v3.3.0 - no-global-assign
+    'no-negated-in-lhs': 'off', // ESLint - v3.3.0 - no-unsafe-negation
+    'no-spaced-func': 'off', // ESLint - v3.3.0 - func-call-spacing
+  },
+  /* eslint-enable sort-keys */
+};