+++ CHANGELOG.md
@@ -19,6 +19,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
   - Fixed typo in module exports for react hooks rules
 - Moved JSX `ecmaFeatures` listing to React config and changed React version from fixed number to `detect`
 - Disabled the `sort-imports` rule as controls are not fine-grained enough, causing more problems than it solved.
+- Explicitly disable deprecated rules
 
 ## [0.1.0](https://github.com/crypto-papers/eslint-config/compare/0.0.1...0.1.0) - 2019-10-15